+++ terraform/test-fixtures/plan-computed-list/main.tf
@@ -1,0 +1,8 @@
+resource "aws_instance" "foo" {
+    num = "2"
+    compute = "list.#"
+}
+
+resource "aws_instance" "bar" {
+    foo = "${aws_instance.foo.list.0.bar}"
+}